samples/contrib/pytorch-samples/requirements.txt
boto3
image
matplotlib==3.4.2
pyarrow
sklearn
transformers
torchtext
webdataset
torchvision
pandas
numpy
s3fs
wget
torch-model-archiver
minio
kfp
tensorboard
torchmetrics
pytorch-lightning==1.4.2
